+++ Hackerthon_Project/learn_edit/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from .models import Blog, Comment
 from .forms import BlogForm, CommentForm
-
+from user.models import User
 
 from django.contrib.auth.decorators import login_required
 # Create your views here.
@@ -21,13 +21,14 @@ def learn_post(request):
             blog.content = form.cleaned_data["content"]
             blog.link = form.cleaned_data["link"]
             blog.create_at = timezone.now()
+            blog.user_id = request.user.id
             blog.save()
             return redirect("learn:learn")
 
 
     else:
         form = BlogForm()
-        return render(request, "learn_edit/learn_post.html", {'form':form})
+    return render(request, "learn_edit/learn_post.html", {'form':form})
 @login_required
 def learn_detail(request, blog_id):
     blog = get_object_or_404(Blog, id=blog_id)
@@ -45,22 +46,34 @@ def learn_detail(request, blog_id):
 @login_required
 def learn_edit(request, blog_id):
     blog = get_object_or_404(Blog, id=blog_id)
+    current_user_id = request.user.id
     if request.method == "POST":
-        form = BlogForm(request.POST, request.FILES, instance=blog)
-        blog = form.save(commit=False)
-        blog.title = form.cleaned_data["title"]
-        blog.content = form.cleaned_data["content"]
-        blog.save()
-        return redirect("learn:learn_detail", blog.id)
+            if form.is_valid():
+                if blog.user.id == current_user_id:
+                        form = BlogForm(request.POST, request.FILES, instance=blog)
+                        blog = form.save(commit=False)
+                        blog.title = form.cleaned_data["title"]
+                        blog.content = form.cleaned_data["content"]
+                        blog.user_id = request.user.id
+                        blog.save()
+                        return redirect("learn:learn_detail", blog.id)
+                else:
+                        return render(request, 'warning.html')
 
     else:
         form = BlogForm(instance=blog)
-        return render(request, "learn_edit/learn_post.html", {'form': form})
+    return render(request, "learn_edit/learn_post.html", {'form': form})
 @login_required
 def learn_delete(request, blog_id):
     blog = get_object_or_404(Blog, id = blog_id)
-    blog.delete()
-    return redirect("learn:learn")
+    current_user_id = request.user.id
+
+    if blog.user.id == current_user_id:
+         blog.delete()
+         return redirect('learn:learn')
+    else:
+
+         return render(request,"warning.html")
 
 """
 댓글 기능
